refactor: BarChart 컴포넌트 폴더 구조 변경, 확장성 있게 수정
