+++ src/components/BarChart/BarChart.stories.tsx
@@ -1,7 +1,10 @@
 import BarChart, { BarChartProps } from "./BarChart.tsx";
 import { Meta } from "@storybook/react";
 import { useState } from "react";
-import { getColors } from "./utils.ts";
+import {
+  getColors,
+  getTitle,
+} from "../../pages/reports/ReportCategoryDetails/utils.ts";
 import { Skeleton } from "@mui/material";
 
 const meta = {
@@ -11,6 +14,7 @@ const meta = {
   args: {
     values: ["used", "predict", "useable"],
     data: [55, 10, 35],
+    titles: ["현재", "예정", "잔액"],
     colors: ["#735BF2", "#DEE0E3", "#F7F7F8"],
     selected: "used",
   },
@@ -28,6 +32,7 @@ export const Example = () => {
   const data = [55, 10, 35];
   const values = ["used", "predict", "useable"];
   const colors = getColors(selected);
+  const titles = values.map((v) => getTitle(v));
 
   return (
     <BarChart
@@ -35,6 +40,7 @@ export const Example = () => {
       data={data}
       colors={colors}
       selected={selected}
+      titles={titles}
       setSelected={setSelected}
     />
   );
